header is over the parallax home is done
