+++ client/src/components/Header.jsx
@@ -15,38 +15,20 @@ export default function Header() {
     navigate(`/search?${searchQuery}`);
   };
 
-  useEffect(() => {
-    const urlParams = new URLSearchParams(location.search);
-    const searchTermFromUrl = urlParams.get('searchTerm');
-    if (searchTermFromUrl) {
-      setSearchTerm(searchTermFromUrl);
-    }
-  }, [location.search]);
+ 
   return (
     // Usa la clase bg-gray-800 para cambiar el color del fondo a un gris oscuro y vibrante
-    <header className="bg-gray-800 text-white shadow-md">
+    <header className="bg-gray-900 text-white opacity-50">
       <div className="flex justify-between items-center max-w-6xl mx-auto p-3">
         <Link to="/">
           <div>
-            <img className="w-40" src={logoOriginal} alt="Logo Habitat" />
+        <img className='w-40' src={logoOriginal} alt="Logo Habitat" />
+
           </div>
-          <h1 className="font-bold text-sm sm:text-xl flex flex-wrap"></h1>
+          <h1 className="font-bold text-sm sm:text-xl flex flex-wrap">
+          </h1>
         </Link>
-        <form
-          onSubmit={handleSubmit}
-          className="bg-white p-3 rounded-lg flex items-center"
-        >
-          <input
-            type="text"
-            placeholder="Buscar..."
-            className="bg-transparent text-gray-500 focus:outline-none w-24 sm:w-64 "
-            value={searchTerm}
-            onChange={(e) => setSearchTerm(e.target.value)}
-          />
-          <button>
-            <FaSearch className="text-gray-600" />
-          </button>
-        </form>
+       
         <ul className="flex gap-4">
           <Link to="/">
             <li className="hidden sm:inline text-slate-700 hover:underline text-white ">
@@ -60,6 +42,7 @@ export default function Header() {
             </Link>
           </li>
           <Link to="/about">
+
             <li className="hidden sm:inline text-gray-100 hover:underline">
               Acerca de
             </li>
@@ -73,10 +56,7 @@ export default function Header() {
               />
             ) : (
               // Usa la clase text-white para cambiar el color del texto a blanco
-              <li className=" text-slate-700 hover:underline text-white">
-                {' '}
-                Sign in
-              </li>
+              <li className=" text-slate-700 hover:underline text-white"> Sign in</li>
             )}
           </Link>
         </ul>
